+++ src/form-integer-field.tsx
@@ -5,15 +5,32 @@ import {useTheme} from './hooks';
 import type {CommonFormFieldProps} from './form';
 import {TextFieldBase} from './text-field-base';
 
-// matches strings like: "x@x.x" (where "x" is any string without spaces)
-const RE_EMAIL = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+const IntegerInput = ({inputRef, value, defaultValue, ...rest}: any) => {
+    const format = (v?: string) => String(v ?? '').replace(/[^\d]/g, '');
 
-export interface FormEmailFieldProps extends CommonFormFieldProps {
+    const handleInput = (e: React.FormEvent<HTMLInputElement>) => {
+        // strip all non numeric characters
+        e.currentTarget.value = format(e.currentTarget.value);
+    };
+    return (
+        <input
+            {...rest}
+            inputMode="numeric" // shows numeric keypad in Chrome for Android
+            pattern="[0-9]*" // shows numeric keypad in iOS
+            onInput={handleInput}
+            ref={inputRef}
+            type="text"
+            value={value === undefined ? undefined : format(value)}
+            defaultValue={defaultValue === undefined ? undefined : format(defaultValue)}
+        />
+    );
+};
+
+export interface FormIntegerFieldProps extends CommonFormFieldProps {
     onChangeValue?: (value: string, rawValue: string) => void;
-    getSuggestions?: (value: string) => Array<string>;
 }
 
-export const FormEmailField: React.FC<FormEmailFieldProps> = ({
+export const FormIntegerField: React.FC<FormIntegerFieldProps> = ({
     disabled,
     error,
     helperText,
@@ -24,7 +41,6 @@ export const FormEmailField: React.FC<FormEmailFieldProps> = ({
     onChangeValue,
     onBlur,
     value,
-    autoComplete = 'email',
     ...rest
 }) => {
     const {texts} = useTheme();
@@ -43,18 +59,14 @@ export const FormEmailField: React.FC<FormEmailFieldProps> = ({
         if (!value) {
             return optional ? '' : texts.formFieldErrorIsMandatory;
         }
-        if (!RE_EMAIL.test(value)) {
-            return texts.formEmailError;
-        }
         return validateProp?.(value, rawValue);
     };
 
-    const processValue = (value: string) => value.replace(/\s/g, '');
+    const processValue = (value: string) => value.trim();
 
     return (
         <TextFieldBase
             {...rest}
-            inputMode="email"
             inputRef={(field) => register({name, field, validate})}
             disabled={disabled || formStatus === 'sending'}
             error={error || !!formErrors[name]}
@@ -75,7 +87,7 @@ export const FormEmailField: React.FC<FormEmailFieldProps> = ({
                 setFormError({name, error: validate?.(values[name], rawValues[name])});
                 onBlur?.(e);
             }}
-            autoComplete={autoComplete}
+            inputComponent={IntegerInput}
         />
     );
 };